refactor: move the storage-backend endpoints out of the closure
Fifth group of #669. StorageBackendInfo, StorageBackendConfig,
StorageBackendTest, StorageBackendMigrate and StorageAzureKeyVaultBackfill move
into modules/api/resources_storage.py, built by
create_storage_resources(api, models, ctx).

Closure complexity 439 -> 372; resources.py 3262 -> 2796 lines. Since #669
started: 559 -> 372, and 4040 -> 2796.

Unlike the earlier groups these register on a namespace they create here rather
than being registered by factory.py from the returned mapping, so they are kept
out of that mapping: adding them would change what create_api_resources hands
back, which test_api_resource_contract pins deliberately. Preserving the
contract matters more here than making the five groups look alike.

Two supporting changes fall out of the move:

test_storage_wiring_consistency reads backend-set literals out of the source by
marker, so it is repointed at the new module. Confirmed still live by dropping
a backend from one literal and watching it fail — a repointed ratchet that no
longer reads anything would otherwise pass forever.

test_manager_lookup_keys_are_real is new, and generalises the defect that hit
the backup group: these endpoints take managers out of the mapping by string
key and answer 503 when the lookup misses, so a wrong key is a total outage
indistinguishable from an unconfigured backend. Rather than a call-through test
per endpoint — which only covers the endpoints someone remembered — it reads
every key out of modules/api and checks it against the mapping factory.py
actually builds, so it covers groups that do not exist yet.

The name rewriting that broke BackupDelete now goes through tokenize rather
than a regex, so it cannot reach inside a string literal.
